+++ src/main/java/io/swagger/api/impl/JobOfferApiServiceImpl.java
@@ -58,7 +58,7 @@ public class JobOfferApiServiceImpl extends JobOfferApiService {
 		//Auth test
 		String[] userPass = getUsernameAndPassword(token);
 		
-		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Integer.valueOf(userPass[0])));
+		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Long.valueOf(userPass[0])));
 		if (restaurantTest == null) 
 			return Response.status(401).entity(new ApiResponseMessage(ApiResponseMessage.ERROR, "unauthorized!")).build();
 		else if (!restaurantTest.getPassword().equals(cryptWithMD5(userPass[1])))
@@ -94,7 +94,7 @@ public class JobOfferApiServiceImpl extends JobOfferApiService {
 		//Auth test
 		String[] userPass = getUsernameAndPassword(token);
 		
-		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Integer.valueOf(userPass[0])));
+		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Long.valueOf(userPass[0])));
 		if (restaurantTest == null) 
 			return Response.status(401).entity(new ApiResponseMessage(ApiResponseMessage.ERROR, "unauthorized!")).build();
 		else if (!restaurantTest.getPassword().equals(cryptWithMD5(userPass[1])))
@@ -119,7 +119,7 @@ public class JobOfferApiServiceImpl extends JobOfferApiService {
 		//Auth test
 		String[] userPass = getUsernameAndPassword(token);
 		
-		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Integer.valueOf(userPass[0])));
+		Restaurant restaurantTest = new RestaurantDao().getRestaurantById(new BigDecimal(Long.valueOf(userPass[0])));
 		if (restaurantTest == null) 
 			return Response.status(401).entity(new ApiResponseMessage(ApiResponseMessage.ERROR, "unauthorized!")).build();
 		else if (!restaurantTest.getPassword().equals(cryptWithMD5(userPass[1])))
